Refactor: Rename js/ directory to src/ (more idiomatic)
Changed the source directory from js/ to src/, which is the standard convention
in modern JavaScript/TypeScript projects. This makes the project structure more
conventional and easier to understand for new contributors.

Changes:
- Renamed js/ directory to src/ using git mv (preserves git history)
- Updated app.js: import path from './js/game' to './src/game'
- Updated tsconfig.json: paths mapping and include patterns
- Updated eslint.config.js: all file patterns and ignore paths
- Updated vitest.config.js: test include patterns
- Updated package.json: lint and lint:fix scripts

All checks passing after rename:
✅ TypeScript: 0 errors (pnpm type-check)
✅ ESLint: 0 errors, 0 warnings (pnpm lint)
✅ Tests: 208/208 passing (pnpm test)
