+++ app.js
@@ -1,7 +1,7 @@
 /*global window:false*/
 import 'normalize.css';
 
-import * as game from './js/game';
+import * as game from './src/game';
 
 const quantumGame = new game.Game();
 quantumGame.htmlReady();
