+++ eslint.config.js
@@ -7,22 +7,22 @@ export default [
   {
     ignores: [
       // Ignore duplicate .js source files (old versions before TypeScript migration)
-      'js/config.js',
-      'js/const.js',
-      'js/tile.js',
-      'js/simulation.js',
-      'js/level.js',
-      'js/winning_status.js',
-      'js/particle/particle.js',
-      'js/tensor/tensor.js',
+      'src/config.js',
+      'src/const.js',
+      'src/tile.js',
+      'src/simulation.js',
+      'src/level.js',
+      'src/winning_status.js',
+      'src/particle/particle.js',
+      'src/tensor/tensor.js',
       // Ignore .js spec files (not in tsconfig)
-      'js/**/*.spec.js',
+      'src/**/*.spec.js',
       // Ignore .d.ts files (type declarations only, not linted by tsconfig)
-      'js/**/*.d.ts',
+      'src/**/*.d.ts',
     ],
   },
   {
-    files: ['js/**/*.{js,ts}'],
+    files: ['src/**/*.{js,ts}'],
     languageOptions: {
       parser: tsparser,
       parserOptions: {
@@ -77,7 +77,7 @@ export default [
     },
   },
   {
-    files: ['js/**/*.js'],
+    files: ['src/**/*.js'],
     rules: {
       '@typescript-eslint/no-explicit-any': 'off',
       '@typescript-eslint/no-unsafe-assignment': 'off',
