+++ package.json
@@ -10,8 +10,8 @@
     "test": "vitest",
     "test:ui": "vitest --ui",
     "type-check": "tsc --noEmit",
-    "lint": "eslint js",
-    "lint:fix": "eslint js --fix"
+    "lint": "eslint src",
+    "lint:fix": "eslint src --fix"
   },
   "dependencies": {
     "d3": "^7.9.0",
