+++ src/bare_board.ts
