+++ src/detection_bar.ts
