+++ src/particle/canvas_particle_animation.ts
