+++ src/particle/particle.js
