+++ src/particle/particle.ts
