+++ src/particle/particle_animation.ts
