+++ src/particle/svg_particle_animation.ts
