+++ src/popup_manager.ts
