+++ src/simulation.js
