+++ src/simulation.ts
