+++ src/sound_service.d.ts
