+++ src/stock.ts
