+++ src/tensor/direction.ts
