+++ src/tensor/full.d.ts
