+++ src/tensor/full.spec.js
