+++ src/tensor/full.ts
