+++ src/tensor/polarization.ts
