+++ src/tensor/tensor.js
