+++ src/tensor/tensor.spec.js
