+++ src/tensor/tensor.ts
