+++ src/tooltip.ts
