+++ src/views/encyclopedia_item_view.ts
