+++ src/views/encyclopedia_selector_view.ts
