+++ src/views/game_view.ts
