+++ src/winning_status.ts
