+++ vitest.config.js
@@ -5,6 +5,6 @@ export default defineConfig({
     globals: true,
     environment: 'jsdom',
     setupFiles: ['./vitest.setup.js'],
-    include: ['js/**/*.spec.js', 'js/**/*.spec.ts'],
+    include: ['src/**/*.spec.js', 'src/**/*.spec.ts'],
   },
 });
